+++ bench/core/computations.hpp
@@ -6,5 +6,5 @@ namespace bench
 {
 double compute_average_frequency(const test_instance &);
 double compute_average_throughput(const test_instance &);
-double compute_iteration_count(const test_instance &);
+long long compute_iteration_count(const test_instance &);
 }
